# Conflicts:
#	Modules/Features/PLLogin/PLLogin/LoginScenes/RememberedLogin/Pin/Coordinator/PLRememberedLoginPinCoordinator.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/RememberedLogin/Pin/Presenter/PLRememberedLoginPinPresenter.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/RememberedLogin/Pin/UseCases/PLRememberedLoginProcessUseCase.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/RememberedLogin/Pin/View/PLRememberedLoginPinViewController.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/SCA/HardwareToken/Presenter/PLHardwareTokenPresenter.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/SCA/SmsAuth/Presenter/PLSmsAuthPresenter.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/TrustedDevice/TrustedDeviceSuccess/Coordinator/PLTrustedDeviceSuccessCoordinator.swift
#	Modules/Features/PLLogin/PLLogin/LoginScenes/TrustedDevice/TrustedDeviceSuccess/Presenter/PLTrustedDeviceSuccessPresenter.swift
#	Project/Santander.xcodeproj/project.pbxproj
#	Project/Santander/Dependencies/AppDependencies.swift
#	Project/Santander/Info.plist
#	Project/Santander/Modifiers/PersonalArea/PersonalInfo/PLPersonalDataModifier.swift
#	santander-one

+++ Modules/Features/BLIK/BLIK/Common/Formatters/AmountFormatter.swift
@@ -5,8 +5,8 @@ import Models
 
 struct AmountFormatter {
     
-    static func amountString(amount: Double, currency: CurrencyType, withAmountSize size: CGFloat) -> NSAttributedString {
-        let moneyDecorator = MoneyDecorator(AmountEntity(value: Decimal(amount), currency: currency),
+    static func amountString(amount: Decimal, currency: CurrencyType, withAmountSize size: CGFloat) -> NSAttributedString {
+        let moneyDecorator = MoneyDecorator(AmountEntity(value: amount, currency: currency),
                                             font: .santander(family: .text, type: .bold, size: size))
         
         let amount =  moneyDecorator.getFormatedCurrency() ?? NSAttributedString(string: "\(amount)")
