+++ Modules/Features/BLIK/BLIK/Features/MobileTransfer/Confirmation/Models/MobileTransfer.swift
@@ -1,8 +1,9 @@
 import Foundation
 import PLUI
+import CoreDomain
 
 public struct MobileTransfer {
-    public let amount: Double?
+    public let amount: Decimal?
     public let title: String?
     public let account: SelectableAccountViewModel
     public let recipientName: String?
